% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_gss_names.R
\name{add_gss_names}
\alias{add_gss_names}
\title{Add a column of GSS names}
\usage{
add_gss_names(
  df_in,
  col_geog = "gss_code",
  col_gss_name = "gss_name",
  gss_date = NA,
  gss_year = NA
)
}
\arguments{
\item{df_in}{A data frame containing gss_codes and data.}

\item{col_geog}{A string. The name of the column which contains gss codes (defaults to
\code{gss_code}).}

\item{col_gss_name}{A string. The name to give the new column containing gss names. 
Defaults to \code{gss_name}. If the column already exists in \code{df_in} it will be
overwritten.}

\item{gss_date}{A date object specifying the point in time that the gss codes 
were/are operational. Only one of gss_date or gss_year can be defined. Defaults 
to \code{NA})}

\item{gss_year}{Numeric or Integer. The year at which the gss codes were/are 
operational. Equivalent to setting gss_date to 31st December of that year. 
Only one of gss_date or gss_year can be defined. Defaults to \code{NA})}
}
\value{
The input dataframe with a column either added or overwritten which contains
the GSS names. The first two columns will be the gss codes and gss names respectively
}
\description{
Adds a column of GSS names to match a column of GSS codes.
}
\details{
Uses the names that were operational for those GSS codes at the given date/year.

Throws an error if there are any GSS codes which were not operational at the 
given date/year.

User must define the name of the column containing GSS codes, give a name
for the new column to contain GSS names, and specify the date or year for the codes and names.

note that 2009 codes run up to 2012 but there was one name change in 2010
and 2013 codes run up to 2019 but there was a name change in 2018.
}
