% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_up_one_geog.R
\name{agg_up_one_geog}
\alias{agg_up_one_geog}
\title{Aggregate numeric data from one geogrpahy level to the next geography level up}
\usage{
agg_up_one_geog(
  df_in,
  col_code = "gss_code",
  col_name = NA,
  col_data = "value",
  fun = "sum",
  gss_date = NA,
  gss_year = NA
)
}
\arguments{
\item{df_in}{A data frame containing gss_codes and data.}

\item{col_code}{A string. The name of the column which contains gss codes (defaults to
\code{gss_code}).}

\item{col_name}{A string. The name of a column containing gss names. If there
is no name column then this should be set to NA. Defaults to NA}

\item{col_data}{A string or character vector. The column(s) that contain the
data to be aggregated. Defaults to 'value'}

\item{fun}{Character. Function to be applied in aggregating data. Either 'sum'
or 'mean'. Default \code{'sum'}.}

\item{gss_date}{A date object specifying the point in time that the gss codes 
were/are operational. Only one of gss_date or gss_year can be defined. Defaults 
to \code{NA})}

\item{gss_year}{Numeric or Integer. The year at which the gss codes were/are 
operational. Equivalent to setting gss_date to 31st December of that year. 
Only one of gss_date or gss_year can be defined. Defaults to \code{NA})}
}
\value{
a dataframe with gss codes and values for one level of geography up
}
\description{
Aggregate numeric data from one geogrpahy level to the next geography level up
}
