% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gss_date.R
\name{get_gss_date}
\alias{get_gss_date}
\title{Get the date range that a set of GSS codes were operational}
\usage{
get_gss_date(df_in, col_geog = "gss_code")
}
\arguments{
\item{df_in}{A data frame containing gss_codes.}

\item{col_geog}{A string. The name of the column which contains gss codes (defaults to
\code{gss_code}).}
}
\value{
A named list with elements called  \code{earliest} and  \code{latest} which gives the 
earliest and latest dates that the gss codeset was operational.
}
\description{
Returns the date range during which all of the gss codes in a dataframe were
operational
}
\details{
Note that 2018 codes will return as 2013 as they are the same as 2013 codes,
only some of the gss names have changed and this doesn't check names.
}
