% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gss_year.R
\name{get_gss_year}
\alias{get_gss_year}
\title{Get the year that a set of GSS codes were operational}
\usage{
get_gss_year(df_in, col_code = "gss_code")
}
\arguments{
\item{df_in}{A data frame containing gss_codes.}

\item{col_code}{A string. The name of the column which contains gss codes (defaults to
\code{gss_code}).}
}
\value{
An integer.  The year of the first December that the 
codes were operational
}
\description{
Returns the earliest year for which all of the gss codes in a dataframe were
operational
}
\details{
A wrapper for get_gss_date which returns the year of the first December that the 
codes were operational
}
