% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_higher_geog.R
\name{add_higher_geog}
\alias{add_higher_geog}
\title{Add a column with GSS codes of a higher geography}
\usage{
add_higher_geog(
  df_in,
  col_code = "gss_code",
  current_geog = c("lad", "region"),
  higher_geog = c("region", "country"),
  gss_date = NA,
  gss_year = NA
)
}
\arguments{
\item{df_in}{A data frame containing GSS codes.}

\item{col_code}{A string. The name of the column which contains gss codes (defaults to
\code{gss_code}).}

\item{current_geog}{A string specifying the geography level of the \code{col_code} column. 
Must be either 'lad' or 'region'.}

\item{higher_geog}{A string specifying the higher geography level to be added.
Must be either 'region' or 'country'.}

\item{gss_date}{A date object specifying the point in time that the gss codes 
were/are operational. One of and only one of gss_date or gss_year must be defined. Defaults 
to \code{NA})}

\item{gss_year}{Numeric or Integer. The year at which the gss codes were/are 
operational. Equivalent to setting gss_date to 31st December of that year. 
One of and only one of gss_date or gss_year must be defined. Defaults to \code{NA})}
}
\value{
The input dataframe with a column added which contains the higher level GSS codes.
}
\description{
Add a new column to a dataframe containing GSS codes where the new column contains 
corresponding GSS codes at a higher geography.
}
\details{
Currently covers LADs, regions, and countries.
 
The name of the new column will be its geography level e.g. 'region'
 
The region code for Wales is coded as the country code ("W92000004) 

Throws an error if there are any GSS codes which were not operational at the 
given date/year.
}
