% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gss_codes.R
\name{check_gss_codes}
\alias{check_gss_codes}
\title{Check that the LA gss codes in a dataframe are consistent with the codes which 
were operational at a given date.}
\usage{
check_gss_codes(
  df_in,
  col_geog = "gss_code",
  gss_date = NA,
  gss_year = NA,
  full_coverage = FALSE,
  include_wales = FALSE
)
}
\arguments{
\item{df_in}{A data frame containing gss_codes and data.}

\item{col_geog}{A string. The name of the column which contains gss codes 
(defaults to \code{gss_code}).}

\item{gss_date}{A date object specifying the point in time that the gss codes 
in the dataframe are being checked against. Only one of gss_date 
or gss_year can be defined. Defaults to \code{NA})}

\item{gss_year}{Numeric or Integer. The year against which to check the gss codes. 
Equivalent to setting gss_date to 31st December of that year. Only one of gss_date 
or gss_year can be defined. Defaults to \code{NA})}

\item{full_coverage}{Logical. If set to TRUE a warning will be given if there are 
codes which are not in df_in but were operational on the date/year given. Defaults to \code{FALSE})}

\item{include_wales}{Logical. If set to TRUE when full_coverage is TRUE, warnings
will be given for missing Welsh codes as well as English ones. Defaults to \code{FALSE})}
}
\value{
Does't return anything
}
\description{
Checks that there are no codes in the data which weren't live on that date,
and can also check for completeness covering England and optionally Wales.
}
\details{
Throws an error if there are codes present which weren't live on the give date,
and warns for any missing codes if full_coverage is set to TRUE.
}
