% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geog_levels.R
\name{get_geog_levels}
\alias{get_geog_levels}
\title{Gets the geography levels present in data}
\usage{
get_geog_levels(df_in, col_code = "gss_code")
}
\arguments{
\item{df_in}{A data frame containing gss_codes.}

\item{col_code}{A string. The name of the column which contains gss codes (defaults to
\code{gss_code}).}
}
\value{
A list of one or more of the following strings: \code{"lad"}, \code{"county"}, \code{"region"} or \code{"country"}
}
\description{
Returns a list of strings describing the geography levels present in the data.
Strings can be "lad", "county", "region" or "country"
}
\details{
Note that Wales will be treated as a region if there are English regions in the 
data but no English country.
}
